requirements-client.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.117.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.60.0
fsspec==2024.6.1
gast==0.5.4
google-pasta==0.2.0
h11==0.16.0
h5py==3.11.0
huggingface-hub==0.35.1
idna==3.10
Jinja2==3.1.4
joblib==1.3.2
keras==3.3.3
kiwisolver==1.4.9
libclang==18.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
opencv-python
opt-einsum==3.3.0
optree==0.11.0
packaging==25.0
pillow==11.0.0
polars==1.33.1
psutil==7.1.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.18.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
rich==13.7.1
safetensors==0.6.2
scikit-learn==1.4.1.post1
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.13.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.4.0
tokenizers==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.37.0
wheel==0.43.0
wrapt==1.16.0
